Update sync_fifo.v
after understanding the full and empty signals are not just for a single address as in that case we can only follow this sequence of operations wr-rd-wr-rd.. or overwrite the already written data when the read enable is 0 but write is happening as it wont update the write pointer as expected. so i changed these full to f and empty to e and introduced fifo_full and fifo_empty which talk about the entire fifo and f,e has no real significance but to not write and read at the same posedge clock as that creates a meta stable condition!